+++ src/Rib/Server.hs
@@ -10,18 +10,20 @@ where
 import Network.Wai.Application.Static (defaultFileServerSettings, staticApp)
 import qualified Network.Wai.Handler.Warp as Warp
 import Relude
+import Rib.Settings
 
 -- | Run a HTTP server to serve a directory of static files
 --
 -- Binds the server to host 127.0.0.1.
 serve ::
+  RibSettings ->
   -- | Port number to bind to
   Int ->
   -- | Directory to serve.
   FilePath ->
   IO ()
-serve port path = do
-  putStrLn $ "[Rib] Serving " <> path <> " at http://" <> host <> ":" <> show port
+serve ribSettings port path = do
+  logStrLn ribSettings $ "[Rib] Serving " <> path <> " at http://" <> host <> ":" <> show port
   Warp.runSettings settings app
   where
     app = staticApp $ defaultFileServerSettings path