+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.9.0.0 (UNRELEASED, DEV)
 
+- Misc changes
+  - #145: CLI arguments have changed
+  - #141: Allow quiet logging (useful when rib is used as a library)
 
 ## 0.8.0.0